Move commands into separate module
